requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain-pinecone 
langchain-openai 
langchain
flask
pypdf
python-dotenv

-e .